[so] squash Library type with so package
