--- pkg/network/so/library.go
@@ -1,12 +1,0 @@
-package so
-
-type libraryKey struct {
-	Pathname       string
-	MountNameSpace ns
-}
-
-type Library struct {
-	libraryKey
-	PidsPath []string
-	HostPath string
-}
